Create README.md
# unspotted
#### this is unspotted web app it links  cleaning service providers to cliets, june 2021
#### By Felix Osambo, fredrick mutsotso and cynthia bella
## Description
This is unspotted web app is a sample web app which links cleaning service providers to clients. whereby you can get a cleaning company or individuals around your area of residence and contact them for your service
## Setup/Installation Requirements
* unspotted
* it contains cleaning companies and individuals who do cleaning services their job and what they have done.
## Known Bugs
unspotted has been built with the following web languages  HTML, CSS and javascript and part of jquary.
## Technologies Used
 the languages and tools you used to create this web-page includes pure HTML, CSS,  javascript jquery and bootsrap.
## Support and contact details
git-hub account: @kodhanjo
email: felix.osambo@students.moringaschool.com
### License
MIT License

Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE. Copyright (c) {2021} Felix Osambo.